src/requirements.txt
streamlit
groq
google-generativeai
openai
gTTS
pandas
